Log In & Redirect, NavBar Fixes, Theme
Login functionality now works although the login form is a little wacky at the moment. The navigation bar now updates based on auth state, showing a logout button if a user is logged in and sign in/create account if not. Fixed themeing issues with text on the homepage. Also updated theme colors.

Signed-off-by: Albert Shih <albertshih3@gmail.com>