Mobile Navigation and Toasts
Introducing the mobile navbar! Now the navbar is fully functional on mobile devices. Also, updated toasts in the app to now display directly from the navbar. This will make it so we do not need to store the toast message in session storage anymore. Toasts should also persist across page changes now.

Signed-off-by: Albert Shih <albertshih3@gmail.com>